Move error_code, bucket and key out from the client crate
