+++ mountpoint-s3-client/src/mock_client.rs
@@ -22,7 +22,7 @@ use time::OffsetDateTime;
 use tracing::trace;
 
 use crate::checksums::crc32c_to_base64;
-use crate::error_metadata::{ErrorMetadata, ProvideErrorMetadata};
+use crate::error_metadata::{ClientErrorMetadata, ProvideErrorMetadata};
 use crate::object_client::{
     Checksum, ChecksumAlgorithm, DeleteObjectError, DeleteObjectResult, ETag, GetBodyPart, GetObjectAttributesError,
     GetObjectAttributesParts, GetObjectAttributesResult, GetObjectError, HeadObjectError, HeadObjectResult,
@@ -516,8 +516,8 @@ impl std::fmt::Display for MockClientError {
 }
 
 impl ProvideErrorMetadata for MockClientError {
-    fn meta(&self) -> &ErrorMetadata {
-        &ErrorMetadata::EMPTY
+    fn meta(&self) -> &ClientErrorMetadata {
+        &ClientErrorMetadata::EMPTY
     }
 }
