+++ mountpoint-s3/src/fs/error.rs
@@ -1,8 +1,8 @@
 //! Utilities for handling errors generated by the `fs` module and mapping them to FUSE errors
 
-use mountpoint_s3_client::error_metadata::{ErrorMetadata, ProvideErrorMetadata};
 use tracing::Level;
 
+use crate::fs::error_metadata::ErrorMetadata;
 use crate::inode::InodeError;
 use crate::upload::UploadWriteError;
 
@@ -168,8 +168,8 @@ impl<E: std::error::Error> ToErrno for UploadWriteError<E> {
     }
 }
 
-impl ProvideErrorMetadata for Error {
-    fn meta(&self) -> &ErrorMetadata {
+impl Error {
+    pub fn meta(&self) -> &ErrorMetadata {
         &self.metadata
     }
 }
