+++ mountpoint-s3/src/inode/readdir.rs
@@ -38,11 +38,11 @@
 use std::cmp::Ordering;
 use std::collections::VecDeque;
 
-use mountpoint_s3_client::error_metadata::MOUNTPOINT_ERROR_CLIENT;
 use mountpoint_s3_client::types::ObjectInfo;
 use mountpoint_s3_client::ObjectClient;
 use tracing::{error, trace, warn};
 
+use crate::fs::error_metadata::MOUNTPOINT_ERROR_CLIENT;
 use crate::sync::{Arc, AsyncMutex, Mutex};
 
 use super::{
